+++ web/src/components/ProcessTimeline2.scss
@@ -39,24 +39,6 @@
   background-color: #f8f9fa;
   font-family: monospace;
 
-  .data-flow-overlay {
-    position: absolute;
-    bottom: 0;
-    height: 100%;
-    overflow: visible;
-
-    .arrow-path {
-      opacity: 0.8;
-    }
-
-    text {
-      font-size: 0.7em;
-    }
-
-    rect {
-      fill: white;
-    }
-  }
 
   .vertical-time-axis {
     width: 70px;
@@ -151,82 +133,4 @@
     }
   }
 
-  .diagram-container {
-    flex: 1;
-    min-width: 0; // Allow it to shrink if needed
-    position: relative;
-    overflow-x: auto;
-    background: white;
-    border: 1px solid #dee2e6;
-    border-radius: 6px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-
-    .function-rectangle {
-      position: absolute;
-      z-index: 1;
-      border-radius: 6px;
-      box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-
-      .function-header {
-
-        padding: 0.75rem 1rem 0 1rem;
-        font-size: 0.9rem;
-        line-height: 1.3;
-
-        .function-name {
-          font-weight: bold;
-        }
-
-        .function-time {
-          font-size: 0.8rem;
-        }
-      }
-
-      .instructions-container {
-        position: relative;
-
-        .instruction-rectangle {
-          z-index: 3;
-          position: absolute;
-          left: 0.5rem;
-          right: 0.5rem;
-          min-height: 50px;
-          box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-          cursor: pointer;
-          display: flex;
-          align-items: center;
-          text-align: center;
-          background: white;
-          border-radius: 6px;
-
-          &:hover {
-            box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
-          }
-
-          .instruction-content {
-            width: 100%;
-            display: flex;
-            flex-direction: column;
-            line-height: 1.2;
-
-            .instruction-label {
-              font-size: 0.8rem;
-              color: #2d3748;
-              font-family: Calibri;
-            }
-
-            .instruction-time {
-              font-size: 0.7rem;
-              color: #6c757d;
-            }
-
-            .instruction-io-info {
-              font-size: 0.65rem;
-              color: #6c757d;
-            }
-          }
-        }
-      }
-    }
-  }
 }
