+++ web/src/components/ProcessTimeline2/FunctionRectangle.scss
@@ -1,0 +1,96 @@
+.function-rectangle {
+  position: absolute;
+  z-index: 1;
+  border-radius: 6px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+
+  .function-header {
+    padding: 0.25rem 1rem 0 1rem;
+    font-size: 0.9rem;
+    line-height: 1.3;
+
+    .function-name {
+      font-weight: bold;
+    }
+
+    .function-time {
+      font-size: 0.8rem;
+    }
+
+    &.left {
+      padding: 0.5rem 1rem 0.5rem 0;
+      writing-mode: vertical-rl;
+      transform: rotate(180deg);
+
+      .function-name {
+        display: flex;
+        flex-direction: column;
+        width: auto;
+        text-align: center;
+        word-break: break-word;
+      }
+
+      .function-time {
+        text-align: center;
+        word-break: break-all;
+      }
+    }
+  }
+
+
+  .instructions-container {
+    position: relative;
+
+    .instruction-rectangle {
+      z-index: 3;
+      position: absolute;
+      left: 0.5rem;
+      right: 0.5rem;
+      min-height: 50px;
+      box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+      cursor: pointer;
+      display: flex;
+      align-items: center;
+      text-align: center;
+      background: white;
+      border-radius: 6px;
+
+      &:hover {
+        box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
+      }
+
+      .instruction-content {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        line-height: 1.2;
+
+        .instruction-label {
+          font-size: 0.8rem;
+          color: #2d3748;
+          font-family: Calibri;
+        }
+
+        //.instruction-time {
+        //  font-size: 0.7rem;
+        //  color: #6c757d;
+        //}
+
+        .instruction-io-info {
+          font-size: 0.65rem;
+          color: #6c757d;
+        }
+      }
+    }
+  }
+
+  // Variant for left header mode in TimelinePerUnit
+  &.with-left-header {
+    display: flex;
+    padding: 0;
+
+    .instructions-container {
+      flex: 1;
+    }
+  }
+}
