+++ web/src/components/ProcessTimeline2/UnitTimeline.scss
@@ -1,0 +1,73 @@
+.timeline-per-unit {
+  width: 100%;
+  height: 100%;
+  background-color: white;
+  border: 1px solid #dee2e6;
+  border-radius: 6px;
+  position: relative;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+  font-family: monospace;
+  overflow-x: auto;
+  overflow-y: hidden;
+
+  .units-header {
+    padding: 1rem;
+    font-weight: bold;
+    border-bottom: 1px solid #dee2e6;
+    background-color: #f8f9fa;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: sticky;
+    top: 0;
+    z-index: 10;
+  }
+
+  .units-container {
+    position: relative;
+    padding: 0 1rem;
+    display: flex;
+    flex-direction: row;
+    gap: 1rem;
+    min-height: 100%;
+  }
+
+  .unit-row {
+    display: flex;
+    flex-direction: column;
+    //border-right: 1px dashed #dee2e6;
+
+    &:last-child {
+      border-right: none;
+    }
+
+    .unit-name {
+      padding: 0.5rem;
+      font-size: 0.8rem;
+      font-weight: bold;
+      text-align: center;
+      word-break: break-all;
+      background-color: #f8f9fa;
+      border-bottom: 1px solid #dee2e6;
+    }
+
+    .unit-column {
+      position: relative;
+      display: inline-block;
+      height: 100%;
+      border: 1px solid #dee2e6;
+      border-radius: 6px;
+      background-color: rgba(255, 255, 255, 0.5);
+      padding: 0.25rem;
+    }
+
+    .unit-timeline-track {
+      position: relative;
+      height: 100%;
+      min-height: 1000px;
+      display: inline-block;
+
+
+    }
+  }
+}
